chore: simplify public interface management
