+++ games/src/interface/position.rs
@@ -1,10 +1,7 @@
 use std::fmt::Display;
 use std::str::FromStr;
 
-use strum::IntoEnumIterator;
-use thiserror::Error;
-
-use super::{BitBoardType, ColoredPieceType, Hash, MoveList, MoveStore, MoveType, SquareType};
+use super::{BitBoardType, ColoredPieceType, Hash, MoveList, MoveStore, MoveType};
 
 /// Position is a generalized interface for board representations of a wide
 /// range of games. It can be used to create game-agnostic software. Tetka
@@ -90,93 +87,4 @@ where
     fn count_moves<const QUIET: bool, const NOISY: bool>(&self) -> usize {
         self.generate_moves::<false, QUIET, NOISY>().len()
     }
-}
-
-/// PositionParseErr represents an error encountered while parsing
-/// the given FEN position field into a valid Position.
-#[derive(Error, Debug)]
-pub enum PiecePlacementParseError {
-    #[error("a jump value was too long and overshot")]
-    JumpTooLong,
-
-    #[error("invalid piece identifier '{0}'")]
-    InvalidPieceIdent(char),
-    #[error("insufficient data to fill the entire {0} file")]
-    FileDataIncomplete(String),
-    #[error("expected {0} ranks, found more")]
-    TooManyRanks(usize),
-}
-
-type Square<P> = <<P as PositionType>::BitBoard as BitBoardType>::Square;
-type File<P> = <Square<P> as SquareType>::File;
-type Rank<P> = <Square<P> as SquareType>::Rank;
-type ColoredPiece<P> = <P as PositionType>::ColoredPiece;
-
-pub fn parse_piece_placement<T: PositionType>(
-    position: &mut T,
-    fen_fragment: &str,
-) -> Result<(), PiecePlacementParseError> {
-    for sq in <<<T as PositionType>::BitBoard as BitBoardType>::Square as IntoEnumIterator>::iter()
-    {
-        position.remove(sq);
-    }
-
-    // Spilt the position spec by the Ranks which are separated by '/'.
-    let ranks: Vec<&str> = fen_fragment.split('/').collect();
-
-    let first_file = File::<T>::iter().next().unwrap();
-
-    let mut file = Ok(first_file);
-    let mut rank = Ok(Rank::<T>::iter().last().unwrap());
-
-    // Iterate over the Ranks in the string spec.
-    for rank_data in ranks {
-        // Rank pointer ran out, but data carried on.
-        if rank.is_err() {
-            return Err(PiecePlacementParseError::TooManyRanks(<<<T as PositionType>::BitBoard as BitBoardType>::Square as SquareType>::Rank::iter().len()));
-        }
-
-        // Iterate over the Square specs in the Rank spec.
-        for data in rank_data.chars() {
-            // Check if a jump spec was too big and we landed on an invalid File.
-            if file.is_err() {
-                return Err(PiecePlacementParseError::JumpTooLong);
-            }
-
-            let file_value = *file.as_ref().unwrap();
-            let rank_value = *rank.as_ref().unwrap();
-            let square = <Square<T>>::new(file_value, rank_value);
-            match data {
-                // Numbers represent jump specs to jump over empty squares.
-                '1'..='8' => {
-                    file = <File<T>>::try_from(file_value.into() + data as u8 - b'1');
-                    if file.is_err() {
-                        return Err(PiecePlacementParseError::JumpTooLong);
-                    }
-                }
-
-                _ => match <ColoredPiece<T>>::from_str(&data.to_string()) {
-                    Ok(piece) => position.insert(square, piece),
-                    Err(_) => return Err(PiecePlacementParseError::InvalidPieceIdent(data)),
-                },
-            }
-
-            // On to the next Square spec in the Rank spec.
-            file = <File<T>>::try_from(file.unwrap().into() + 1);
-        }
-
-        // After rank data runs out, file pointer should be
-        // at the last file, i.e, rank is completely filled.
-        if let Ok(file) = file {
-            return Err(PiecePlacementParseError::FileDataIncomplete(
-                file.to_string(),
-            ));
-        }
-
-        // Switch rank pointer and reset file pointer.
-        rank = <Rank<T>>::try_from((rank.unwrap().into()).wrapping_sub(1));
-        file = Ok(first_file);
-    }
-
-    Ok(())
 }
